Refactorings: * Extracted LedgerEndPoller out of ReadOnlyLedger * Moves DispatcherLagMeter in its own file
