+++ ledger/participant-integration-api/src/main/scala/platform/index/ReadOnlyLedger.scala
@@ -30,7 +30,7 @@ import com.daml.platform.store.entries.{ConfigurationEntry, PackageLedgerEntry, 
 import scala.concurrent.Future
 
 /** Defines all the functionalities a Ledger needs to provide */
-private[platform] trait ReadOnlyLedger extends ReportsHealth with AutoCloseable {
+private[platform] trait ReadOnlyLedger extends ReportsHealth {
 
   def ledgerId: LedgerId
